examples/requirements.txt
Pillow==10.3.0
av==12.0.0
httpx==0.27.0
playsound==1.3.0
pygobject==3.48.2
